AI.long-running-agents/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "long-running-agents"
version = "0.1.0"
description = "A dual-agent framework for long-running AI agents"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.39.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "gitpython>=3.1.0",
    "click>=8.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
web = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "websockets>=12.0.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
long-agents = "src.__main__:run"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
